requirements.txt
numpy>=1.24.0
Flask==3.0.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
web3==6.11.3
python-dotenv==1.0.0
